% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi-package.R
\docType{package}
\name{gurobi-package}
\alias{gurobi-package}
\alias{gurobi_package}
\title{Gurobi R package}
\description{
This section documents the Gurobi R interface. For those of you who
are not familiar with R, it is an environment for doing numerical
computing, a free language for statistical computing. Please visit
the \href{https://www.r-project.org}{R website} for more
information. This manual begins with a quick overview of the
methods provided by our R API.  It then continues with a
comprehensive presentation of all of the available methods,their
arguments, and their return values.  If you are new to the Gurobi
Optimizer, we suggest that you start with the \href{https://www.gurobi.com/documentation/quickstart.html}{Quick Start Guide},
\href{https://www.gurobi.com/documentation/9.0/examples/index.html}{Example Tour}.
These documents provide concrete examples of how to use the methods
described here.
}
\section{Notes}{

A quick note for new users: the convention in math programming is
that variables are non-negative unless specified otherwise.  You'll
need to explicitly set lower bounds if you want variables to be
able to take negative values.
}

\author{
Gurobi Optimization
}
\keyword{optimize}
