% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi.R
\name{gurobi_iis}
\alias{gurobi_iis}
\title{Compute an Irreducible Inconsistent Subsystem (IIS).}
\usage{
gurobi_iis(model, params = NULL, env = NULL)
}
\arguments{
\item{model}{The model list must contain a valid Gurobi model. See
\code{\link{gurobi_model}} for more information.}

\item{params}{The params \code{list}, when provided, contains a list of
modified Gurobi parameters. See the \code{\link{gurobi_params}}
documentation for more details.}

\item{env}{The env list, when provided, allows you to use Gurobi Compute
Server or Gurobi Instant Cloud. See \code{\link{gurobi_env}} for more
information.}
}
\value{
The \code{gurobi_iis()} function returns a \code{list}, with various results
stored in its named components. The specific results that are available
depend on the type of model.

The returned \code{list} will always contain the following \code{named
components}: \itemize{ \item\link{minimal} A logical scalar that indicates
whether the computed IIS is minimal. It will normally be true, but it may be
false if the IIS computation was stopped early (due to a time limit or a
user interrupt). \item\link{Arows} c("A logical vector that indicates wheather a
#1 appears in the computed IIS.", "linear constraint")A logical vector that
indicates wheather a linear constraint appears in the computed IIS.
\item\link{lb} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "lower bound")A logical vector that indicates wheather a
lower bound appears in the computed IIS. \item\link{ub} c("A logical vector that
indicates wheather a #1 appears in the computed IIS.", "upper bound")A
logical vector that indicates wheather a upper bound appears in the computed
IIS. }

If your model contains general constraints, the returned \code{list} will
also contain the following \code{named components}: \itemize{ \item
\link{genconmax} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general MAX constraint")A logical vector that indicates
wheather a general MAX constraint appears in the computed IIS. \item
\link{genconmin} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general MIN constraint")A logical vector that indicates
wheather a general MIN constraint appears in the computed IIS. \item
\link{genconand} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general AND constraint")A logical vector that indicates
wheather a general AND constraint appears in the computed IIS. \item
\link{genconor} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general OR constraint")A logical vector that indicates
wheather a general OR constraint appears in the computed IIS. \item
\link{genconabs} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general ABS constraint")A logical vector that indicates
wheather a general ABS constraint appears in the computed IIS. \item
\link{genconind} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "general INDICATOR constraint")A logical vector that
indicates wheather a general INDICATOR constraint appears in the computed
IIS. \item \link{genconpwl} c("A logical vector that indicates wheather a #1
appears in the computed IIS.", "general piecewise-linear function
constraint")A logical vector that indicates wheather a general
piecewise-linear function constraint appears in the computed IIS. \item
\link{genconpoly} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "polynomial function constraint")A logical vector that
indicates wheather a polynomial function constraint appears in the computed
IIS. \item \link{genconexp} c("A logical vector that indicates wheather a #1
appears in the computed IIS.", "natural exponential function constraint")A
logical vector that indicates wheather a natural exponential function
constraint appears in the computed IIS. \item \link{genconexpa} c("A logical
vector that indicates wheather a #1 appears in the computed IIS.",
"exponential function constraint")A logical vector that indicates wheather a
exponential function constraint appears in the computed IIS. \item
\link{genconlog} c("A logical vector that indicates wheather a #1 appears in the
computed IIS.", "natural logarithmic function constraint")A logical vector
that indicates wheather a natural logarithmic function constraint appears in
the computed IIS. \item \link{genconloga} c("A logical vector that indicates
wheather a #1 appears in the computed IIS.", "logarithmic function
constraint")A logical vector that indicates wheather a logarithmic function
constraint appears in the computed IIS. \item \link{genconpow} c("A logical
vector that indicates wheather a #1 appears in the computed IIS.", "power
function constraint")A logical vector that indicates wheather a power
function constraint appears in the computed IIS. \item \link{genconsin} c("A
logical vector that indicates wheather a #1 appears in the computed IIS.",
"SIN function constraint")A logical vector that indicates wheather a SIN
function constraint appears in the computed IIS. \item \link{genconcos} c("A
logical vector that indicates wheather a #1 appears in the computed IIS.",
"COS function constraint")A logical vector that indicates wheather a COS
function constraint appears in the computed IIS. \item \link{gencontan} c("A
logical vector that indicates wheather a #1 appears in the computed IIS.",
"TAN function constraint")A logical vector that indicates wheather a TAN
function constraint appears in the computed IIS. }

If your model contains SOS constraints, the returned \code{list} will also
contain the following \code{named component}: \itemize{ \item\link{sos} A logical
vector that indicates whether an SOS constraint appears in the computed IIS
}

If your model contains quadratic constraints, the returned \code{list} will
also contain the following \code{named component}: \itemize{ \item\link{quadcon}
c("A logical vector that indicates wheather a #1 appears in the computed
IIS.", "quadratic constraint")A logical vector that indicates wheather a
quadratic constraint appears in the computed IIS. }
}
\description{
An IIS is a subset of the constraints and variable bounds with the following
properties: \itemize{ \item the subsystem represented by the IIS is
infeasible, and \item if any of the constraints or bounds of the IIS is
removed, the subsystem becomes feasible. } Note that an infeasible model may
have multiple IISs. The one returned by Gurobi is not necessarily the one
with minimum cardinality; there may exist others with fewer constraints or
bounds.  If an IIS computation is interrupted before completion, Gurobi will
return the smallest IIS found to that point.
}
\details{
You can obtain information about the outcome of the IIS computation from the
returned IIS result (described below). Note that this method can be used to
compute IISs for both continuous and MIP models.
}
\examples{
\%
model <-  gurobi_read('examples/data/klein1.mps')
iis <-  gurobi_iis(model)

}
\references{
Gurobi Optimization (\url{http://www.gurobi.com}).

Gurobi Optimizer Reference Manual
(\url{http://www.gurobi.com/documentation/9.0/refman/index.html}).
}
\seealso{
The \code{\link{gurobi_package}} documentation contains general
information on the Gurobi R interface. The \code{\link{gurobi_problem}}
documentation has a detailed description of the types of models that gurobi
is able to optimize. The \code{\link{gurobi_usage}} documentation has an
overview on the usage of the Gurobi package.

For a detailed description of variables used as input and output of gurobi
functions see \code{\link{gurobi_model}}, \code{\link{gurobi_params}} and
\code{\link{gurobi_env}}.

The functions provided by the Gurobi package are: \code{\link{gurobi}},
\code{\link{gurobi_iis}}, \code{\link{gurobi_feasrelax}},
\code{\link{gurobi_relax}}, \code{\link{gurobi_read}} and
\code{\link{gurobi_write}}.
}
\author{
Gurobi Optimization
}
\keyword{optimize}
